requirements.txt
discord.py==2.0.0a3407+gfc51736b
python-dotenv==0.15.0
pytz==2021.1
emoji==1.2.0
